Give ScreenPainter another function
I am making a program that draws circles and I noticed there were no functions for that but I could make it happen by using drawArc() the right way so I figured it would be a good idea to add a drawCircle() function. This new function was crafted using the preexisting drawArc() function. I have tested it and it is much more practical than using drawArc() when working with pure circles.